requirements.txt
python-dotenv~=1.0.1
ollama~=0.4.7